% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anderson_index.R
\name{anderson_index}
\alias{anderson_index}
\title{Calculate Anderson Index}
\usage{
anderson_index(xmat, revcols = NULL, sgroup = rep(TRUE, nrow(xmat)))
}
\arguments{
\item{xmat}{A matrix of outcomes.}

\item{revcols}{An optional vector indicating columns to reverse signs for "better" outcomes.}

\item{sgroup}{A logical vector indicating the control group for standardization.}
}
\value{
A list containing the index and weights.
}
\description{
Creates an index as a weighted average of outcomes, where weights are based on the inverse covariance matrix of outcomes.
}
\examples{
xmat <- matrix(rnorm(20), nrow = 5)
Anderson_Index(xmat)
}
